add non-fun type to lambda syntax
The non-fun type encompasses any values that are not functions. If a reduction encounters a value of type non-fun being applied to something, it will raise an error (this is the err-app case). Before making this change, we were using a side-condition that checked if the value being applied was a function in Λ-eval, which caused issues if it was a function in ΛC-eval but not a function in Λ-eval.